requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.15
django-admin-rangefilter==0.8.7
django-autocomplete-light==3.9.4
django-autocompletefilter==0.0.8
django-clearcache==1.2.1
django-cors-headers==3.13.0
django-crum==0.7.9
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dnspython==2.2.1
email-validator==1.2.1
fhir.resources==6.4.0
flower==1.2.0
humanize==4.3.0
idna==3.3
kombu==5.2.4
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pydantic==1.10.2
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
gunicorn==20.0.4
drf-spectacular==0.26.5
Hashids==1.3.1
django-select2==8.1.2
django-import-export==2.9.0